+++ common/src/main/java/com/railwayteam/railways/base/data/recipe/RailwaysRecipeProvider.java
@@ -2,6 +2,7 @@ package com.railwayteam.railways.base.data.recipe;
 
 import com.railwayteam.railways.Railways;
 import com.railwayteam.railways.multiloader.CommonTags;
+import com.railwayteam.railways.registry.CRBlocks;
 import com.simibubi.create.AllBlocks;
 import com.simibubi.create.AllItems;
 import net.minecraft.core.Registry;
@@ -11,6 +12,7 @@ import net.minecraft.data.recipes.RecipeProvider;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.ItemTags;
 import net.minecraft.tags.TagKey;
+import net.minecraft.world.item.DyeColor;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.level.ItemLike;
@@ -158,6 +160,30 @@ public abstract class RailwaysRecipeProvider extends RecipeProvider {
 
     public static ItemLike smallCog() {
       return AllBlocks.COGWHEEL.get();
+    }
+
+    public static ItemLike ironBlock() {
+      return Blocks.IRON_BLOCK;
+    }
+
+    public static TagKey<Item> dye(@NotNull DyeColor color) {
+      return CommonTags.DYES.get(color).tag;
+    }
+
+    public static TagKey<Item> brassIngot() {
+      return CommonTags.BRASS_INGOTS.tag;
+    }
+
+    public static ItemLike shaft() {
+      return AllBlocks.SHAFT.get();
+    }
+
+    public static ItemLike linkPin() {
+      return CRBlocks.LINK_AND_PIN.get();
+    }
+
+    public static ItemLike headstock() {
+      return CRBlocks.HEADSTOCK.get();
     }
   }
 }