+++ common/src/main/java/com/railwayteam/railways/content/extended_sliding_doors/SlidingDoorMode.java
@@ -107,6 +107,6 @@ public enum SlidingDoorMode implements INamedIconOptions {
     }
 
     public interface IHasDoorMode {
-        SlidingDoorMode getSlidingDoorMode();
+        SlidingDoorMode snr$getSlidingDoorMode();
     }
 }