+++ common/src/main/java/com/railwayteam/railways/mixin/MixinSlidingDoorBlock.java
@@ -6,6 +6,7 @@ import com.simibubi.create.content.decoration.slidingDoor.SlidingDoorBlockEntity
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
@@ -14,6 +15,7 @@ import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.properties.DoubleBlockHalf;
 import net.minecraft.world.phys.BlockHitResult;
+import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -32,7 +34,7 @@ public abstract class MixinSlidingDoorBlock {
                                                        boolean lower, boolean isPowered, SlidingDoorBlockEntity be) {
         if (be == null)
             return;
-        if (!((SlidingDoorMode.IHasDoorMode) be).getSlidingDoorMode().canOpenSpecially()) {
+        if (!((SlidingDoorMode.IHasDoorMode) be).snr$getSlidingDoorMode().canOpenSpecially()) {
             ci.cancel();
         }
     }
@@ -53,7 +55,17 @@ public abstract class MixinSlidingDoorBlock {
                                                   CallbackInfoReturnable<InteractionResult> cir) {
         boolean lower = pState.getValue(SlidingDoorBlock.HALF) == DoubleBlockHalf.LOWER;
         BlockEntity be = pLevel.getBlockEntity(lower ? pPos : pPos.below());
-        if (be instanceof SlidingDoorMode.IHasDoorMode doorMode && !doorMode.getSlidingDoorMode().canOpenManually())
+        if (be instanceof SlidingDoorMode.IHasDoorMode doorMode && !doorMode.snr$getSlidingDoorMode().canOpenManually())
             cir.setReturnValue(InteractionResult.FAIL);
+    }
+
+    @Inject(method = "setOpen", at = @At("HEAD"), cancellable = true, remap = true)
+    private void preventSpecialDoorEntityOpen(@Nullable Entity entity, Level level, BlockState state, BlockPos pos, boolean _open, CallbackInfo ci) {
+        if (entity != null) {
+            boolean lower = state.getValue(SlidingDoorBlock.HALF) == DoubleBlockHalf.LOWER;
+            BlockEntity be = level.getBlockEntity(lower ? pos : pos.below());
+            if (be instanceof SlidingDoorMode.IHasDoorMode doorMode && !doorMode.snr$getSlidingDoorMode().canOpenManually())
+                ci.cancel();
+        }
     }
 }