+++ common/src/main/java/com/railwayteam/railways/registry/CRTags.java
@@ -1,6 +1,7 @@
 package com.railwayteam.railways.registry;
 
 import com.railwayteam.railways.Railways;
+import com.railwayteam.railways.util.TextUtils;
 import com.simibubi.create.foundation.utility.Lang;
 import net.minecraft.core.Registry;
 import net.minecraft.resources.ResourceLocation;
@@ -10,6 +11,8 @@ import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
+
+import java.util.function.BiConsumer;
 
 import static com.railwayteam.railways.registry.CRTags.NameSpace.MOD;
 
@@ -41,7 +44,9 @@ public class CRTags {
   public enum AllBlockTags {
     SEMAPHORE_POLES(MOD,MOD.optionalDefault,false),
     TRACK_CASING_BLACKLIST(MOD,MOD.optionalDefault,false),
-    CONDUCTOR_SPY_USABLE(MOD, MOD.optionalDefault,false) // so other mods / datapacks can make more blocks usable for conductor spies
+    CONDUCTOR_SPY_USABLE(MOD, MOD.optionalDefault,false), // so other mods / datapacks can make more blocks usable for conductor spies
+    LOCOMETAL,
+    LOCOMETAL_BOILERS
     ;
 
     public final TagKey<Block> tag;
@@ -138,5 +143,20 @@ public class CRTags {
   public static void register() {
     AllBlockTags.register();
     AllItemTags.register();
+  }
+
+  public static void provideLangEntries(BiConsumer<String, String> consumer) {
+    for (AllBlockTags blockTag : AllBlockTags.values()) {
+      ResourceLocation loc = blockTag.tag.location();
+      consumer.accept("tag.block." + loc.getNamespace() + "." + loc.getPath().replace('/', '.'),
+          TextUtils.titleCaseConversion(blockTag.name()).replace('_', ' '));
+    }
+
+    for (AllItemTags itemTag : AllItemTags.values()) {
+      ResourceLocation loc = itemTag.tag.location();
+      consumer.accept("tag.item." + loc.getNamespace() + "." + loc.getPath().replace('/', '.'),
+          TextUtils.titleCaseConversion(itemTag.name()).replace('_', ' '));
+    }
+    consumer.accept("tag.item.forge.string", "String");
   }
 }